Small bugfix concerning "roundtrip_enabled" settings
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/extbase_kickstarter/branches/extension_editing@38735 735d13b6-9817-0410-8766-e36946ffe9aa

+++ Classes/Service/CodeGenerator.php
@@ -76,7 +76,6 @@ class Tx_ExtbaseKickstarter_Service_CodeGenerator implements t3lib_Singleton {
 	public function build(Tx_ExtbaseKickstarter_Domain_Model_Extension $extension) {
 		$this->extension = $extension;
 		$this->classBuilder->injectExtension($extension);
-		$this->roundTripService->injectExtension($extension);
 		if($this->config['settings']['enableRoundtrip']){
 			$this->roundTripService->injectExtension($extension);
 		}
@@ -133,7 +132,7 @@ class Tx_ExtbaseKickstarter_Service_CodeGenerator implements t3lib_Singleton {
 			$typoscriptDirectory = $extensionDirectory . 'Configuration/TypoScript/';
 			$fileContents = $this->generateTyposcriptSetup($extension);
 			$targetFile = $typoscriptDirectory . 'setup.txt';
-			if(!file_exists($targetFile) || $this->roundTripService->getOverWriteSetting($targetFile) < 2){
+			if(!file_exists($targetFile) || ($this->config['settings']['enableRoundtrip'] && $this->roundTripService->getOverWriteSetting($targetFile) < 2)){
 				t3lib_div::writeFile($typoscriptDirectory . 'setup.txt', $fileContents);
 			}
 		} catch (Exception $e) {
